requirements.txt
streamlit
tensorflow
numpy
pillow
reportlab
openpyxl
python-docx
pandas
scikit-learn
